refactoring: move sso to the sso.py & add more type hints
